app/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
firebase-admin==6.5.0
httpx==0.27.2
pydantic[email]==2.9.2
google-cloud-firestore
openai==1.45.0
stripe
redis==4.10.2
aioredis==2.0.1
pytest==7.4.2
pytest-asyncio==0.22.0
